% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controllabilite_Matix.R
\name{calculate_matrix_controllability}
\alias{calculate_matrix_controllability}
\title{Calculate the Controllability Matrix}
\usage{
calculate_matrix_controllability(A, B)
}
\arguments{
\item{A}{The state matrix.}

\item{B}{The Control Matrix}
}
\value{
A controllability matrix.
}
\description{
This function calculates the controllability matrix for the given system.
}
\examples{
\dontrun{
A <- matrix(c(0, 1, -2, -3), nrow = 2, byrow = TRUE)
B <- matrix(c(0, 1), nrow = 2, byrow = TRUE)
calculer_matrice_controllabilite(A, B)
}
}
