% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kalman_criterion.R
\name{kalman_criterion}
\alias{kalman_criterion}
\title{Calculate and verify controllability}
\usage{
kalman_criterion()
}
\value{
A message indicating whether the system is controllable.
}
\description{
This function reads the state and control matrices, calculates the controllability matrix, and checks whether the system is controllable.
}
\examples{
\dontrun{
kalman_criterion()
}
}
