% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterKalman.R
\name{kalman_filter}
\alias{kalman_filter}
\title{Title
This function applies a Kalman filter to estimate the state of a linear system over time.}
\usage{
kalman_filter(observations, A, C, Q, R, x0, P0)
}
\arguments{
\item{observations}{A numerical vector representing observations over time.}

\item{A}{A State transition coefficient}

\item{C}{A Coefficient of observation}

\item{Q}{The variance of the process noise, which represents the uncertainty of the state prediction.}

\item{R}{Variance of the observational noise , which represents the uncertainty in the observed measurements.}

\item{x0}{Initial condition estimate}

\item{P0}{Initial Covariance of State Estimation}
}
\value{
A list
}
\description{
Title
This function applies a Kalman filter to estimate the state of a linear system over time.
}
\examples{
observations <- rnorm(50, mean = 100, sd = 10)
A <- 1
C <- 1
Q <- 1
R <- 1
x0 <- 100
P0 <- 10

# Application du filtre de Kalman
result <- kalman_filter(observations, A, C, Q, R, x0, P0)

# Affichage des résultats
print(result$x_hat)       # Estimations de l'état
print(result$residuals)   # Résidus
print(result$P)           # Covariance
}
