% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Control_Matrix.R
\name{read_control_matrix}
\alias{read_control_matrix}
\title{Control Matrix}
\usage{
read_control_matrix(nb_states)
}
\arguments{
\item{nb_states}{}
}
\value{
a matrix representing the control matrix.
}
\description{
Control Matrix
}
\examples{
\dontrun{
lire_matrice_controle(3)
}
}
