% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StateMatrix.R
\name{read_state_matrix}
\alias{read_state_matrix}
\title{State Matrix}
\usage{
read_state_matrix()
}
\value{
A matrix representing the states of the system.
}
\description{
State Matrix
}
\examples{
\dontrun{
lire_matrice_etat()
}
}
