% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verifyControllability.R
\name{verify_controllability}
\alias{verify_controllability}
\title{Verify system controllability}
\usage{
verify_controllability(A, Control)
}
\arguments{
\item{A}{The state matrix.}

\item{Control}{The controllability matrix.}
}
\value{
TRUE if the system is controllable, otherwise FALSE.
}
\description{
This function checks whether the system is controllable.
}
\examples{
\dontrun{
A <- matrix(c(0, 1, -2, -3), nrow = 2, byrow = TRUE)
C <- matrix(c(0, 1, 1, -3), nrow = 2, byrow = TRUE)
verifier_controllabilite(A, C)
}
}
